style: add tooltip & delete style for cloud cover
